refactor: replaces Token type throughout
